グラフ表示をまともに戻した
git-svn-id: file:///home/yoshi/.svnrepos/vsd/trunk@463 177d47f1-6453-0410-9475-f12f96c2b1a2

+++ vsd_filter/CVsdFilter.h
@@ -152,7 +152,7 @@ class CVsdFilter {
 		CVsdFont &Font,
 		tRABY uColor,
 		CVsdLog& Log,
-		double ( CVsdLog::*GetDataFunc )( int ),
+		double ( CVsdLog::*GetDataFunc )( double ),
 		double dMaxVal
 	);
